Adds Downtime Awareness
This adds a Downtime model and relevant operations so that
the dashboard knows when a lab is down for maintenance and
can act accordingly.

This change doesn't modify the front end at all,
but it does pass relevant downtime info to the
templates so that they can be updated in a future change.

Change-Id: Idb88b15838b949f352f11a31a1fce9749d283d28
Signed-off-by: Parker Berberian <pberberian@iol.unh.edu>