devel: use mounts for development
This removes the need to rebuild the containers every time for simple
changes during development.
This does not include python dependencies since those are installed
globally in the container. (I would have done the same for JS
dependencies, but it's used in the static files.)

Removed the background flag for running in dev mode when using make.
It's easier to kill the servers (^C) and view logs of the server without
needing to call separate commands later.

Nginx is disabled in dev mode since the server and static files are
handled by Django instead.

Update readme to reflect upon changes made.

Signed-off-by: Jeremy Plsek <jeremyplsek@gmail.com>
Change-Id: I7888ca89021fca313e1043a7f94b5e1b7e12498c